desluseste-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyMuPDF
aiohttp
python-dotenv
requests